[rabbitmq] improve RabbitMQNotReady alert for deployments
The critical alert should not fire if any rabbitmq is not ready,
it should fire if no rabbitmq pods are ready.
Which is is difference, because there may be a NotReady pod still
lingering around next to a working Ready pod.
We observed this with a pod stuck in Terminating due a node not being
safely drained.

In case of a statefulset we keep the old setting: we want to be alerted
if any pod is not ready.